Move TagService to another file. Add docstring for TagService.
