+++ src/transcriber_service/services/__init__.py
@@ -2,5 +2,6 @@ from .auth import AuthService
 from .storage_service import StorageService
 from .processing import WhisperProcessor
 from .audio_service import AudioService
+from .tag_service import TagService
 
-__all__ = ['WhisperProcessor', 'AuthService', 'StorageService', 'AudioService']
+__all__ = ['WhisperProcessor', 'AuthService', 'StorageService', 'AudioService', 'TagService']
